Move contentful service into networkCalls directory and update imports
