+++ __test__/helpers/sortEventsByTime/sortEventsByTime.fixture.ts
@@ -1,4 +1,4 @@
-import { ParsedEvent } from "@/app/contentful/contentfulServices.types";
+import { ParsedEvent } from "@/app/networkCalls/contentful/contentfulServices.types";
 import { BLOCKS } from "@contentful/rich-text-types";
 
 export const eventsFixture: ParsedEvent[] = [
