+++ src/app/components/CheckoutDialog/CheckoutDialog.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { ParsedEvent } from "@/app/contentful/contentfulServices.types";
+import { ParsedEvent } from "@/app/networkCalls/contentful/contentfulServices.types";
 import {
   Dialog,
   DialogContent,
