+++ src/app/networkCalls/contentful/contentfulService.ts
