+++ src/app/networkCalls/contentful/graphQL/queries/photoAlbum.ts
