+++ src/app/networkCalls/contentful/parsers/parseEvents.ts
