+++ src/app/networkCalls/contentful/parsers/parsePhotoGallery.ts
